Fix $?CLASS et al leaking outside the package's scope
In addition to adding generated lexicals for $?CLASS, ::?ROLE, etc. to a
package's block, we need to declare those lexicals at BEGIN time, so they are
available to traits, signatures, the parser etc. However doing so in
PERFORM-BEGIN would get them added to the surrounding scope, thus making them
still visible after the class body was finished.

We need to turn the package itself into a lexical scope, to get those lexicals
visible as much as possible, but not more than they should be.

This neccessitates a change to where we add generated protos for multis as
well. I don't really know why though. Probably two bugs cancelling each other
out. In the end I have never been comfortable with the previous implementation
anyway as it abused attach targets for finding a lexical scope.